+++ pdp_graphql_client_python/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""Top-level package for PDP GraphQL client Python."""
-
-__author__ = """Davide Heller"""
-__email__ = 'davide.heller@srf.ch'
-__version__ = '0.1.0'