+++ HISTORY.rst
@@ -2,7 +2,33 @@
 History
 =======
 
-0.1.0 (2021-01-27)
+0.4.0 (2021-03-24)
+------------------
+
+* Removed the dependency on environment variables.
+* Moved to class design.
+
+
+0.3.3 (2021-03-24)
+------------------
+
+* Nothing changed, just testing the release pipeline.
+
+
+0.3.2 (2021-03-24)
+------------------
+
+* Removed the dependency on Click
+
+
+0.3.1 (2021-03-24)
+------------------
+
+* Fixed typo in README.rst
+* Added missing dependency to sgqlc
+
+
+0.3.0 (2021-03-23)
 ------------------
 
 * First release on PyPI.